Move shims into their own file
Circular dependencies gonna getcha
