+++ mautrix_instagram/db/__init__.py
@@ -6,24 +6,7 @@ from .puppet import Puppet
 from .reaction import Reaction
 from .upgrade import upgrade_table
 from .user import User
-
-try:
-    import asyncpg
-
-    UniqueError = asyncpg.UniqueViolationError
-    Record = asyncpg.Record
-except ImportError:
-    pass
-try:
-    import sqlite3
-
-    UniqueError = sqlite3.IntegrityError
-    Record = sqlite3.Row
-except ImportError:
-    pass
-
-if UniqueError is None:
-    raise ImportError("Must require either asyncpg or aiosqlite!")
+from .shim import UniqueError
 
 
 def init(db: Database) -> None:
