+++ mautrix_instagram/db/portal.py
@@ -22,7 +22,7 @@ from attr import dataclass
 from mautrix.types import ContentURI, RoomID, UserID
 from mautrix.util.async_db import Database
 
-from . import Record
+from .shim import Record
 
 fake_db = Database.create("") if TYPE_CHECKING else None
